# app/backend/openapi.yaml
openapi: 3.0.1
info:
  title: Gigator API
  version: 0.1.0
  description: Minimal OpenAPI skeleton for core Gigator backend endpoints (opportunities, threads, compose, auth, AI).
servers:
  - url: http://localhost:3000
    description: Local dev
paths:
...
  /api/contacts/{id}/venues:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List venues linked to a contact
      responses:
        '200':
          description: Array of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
    post:
      summary: Link contact to a venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - venue_id
              properties:
                venue_id:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "not_found"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
  /api/contacts/{id}/venues/{venue_id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: venue_id
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Unlink contact from a venue
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
  /api/venues/{id}/contacts:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List contacts linked to a venue
      responses:
        '200':
          description: Array of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "server_error"
...
  /api/admin/members:
    get:
      summary: List members (app allowlist)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of allowlisted members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminMember'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "forbidden"

    post:
      summary: Invite member (add/reactivate allowlist entry)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Upserted allowlist entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMember'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/members/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Activate/deactivate member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - is_active
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: Updated allowlist entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMember'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "not_found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      description: Generic error response payload.
      required:
        - error
      properties:
        error:
          type: string
          description: Stable error identifier.
        message:
          type: string
          description: Human-readable error message (safe to expose).
...
    AdminMember:
      type: object
      required:
        - id
        - email
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
...
    Contact:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
    Venue:
      type: object
      required:
        - id
        - name
        - city
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
...